internal/pwru: fix printing of function names on arm64
It seems the event address is already correct for the addr2name lookup
on arm64, not sure why the decrement by 1 is needed on x86. Fix this up
and also make the addr2name map lookup more robust by checking the
returned ksym before dereferencing it.

This allows events to be reported properly on arm64.

Signed-off-by: Tobias Klauser <tobias@cilium.io>