Drop dependency on github.com/cilium/cilium
Currently, the cilium Go module is only pulled in for the
github.com/cilium/cilium/pkg/byteorder package. Only two functions out
of that package are used in pwru. Cilium pulls in lots of
transitive dependencies and also requires keeping the replace directives
in go.mod in sync manually. To avoid that overhead, drop that dependecy
and copy the relevant parts of pkg/byteorder to a local package. Also
pull in the changes from https://github.com/cilium/cilium/pull/18606 for
some performance improvements in byte swapping operations.

Signed-off-by: Tobias Klauser <tobias@cilium.io>